+++ src/MailchimpCommerce.php
@@ -250,6 +250,7 @@ class MailchimpCommerce extends Plugin
 			$event->rules['mailchimp-commerce/list'] = 'mailchimp-commerce/cp/list';
 			$event->rules['mailchimp-commerce/mappings'] = 'mailchimp-commerce/cp/mappings';
 			$event->rules['mailchimp-commerce/settings'] = 'mailchimp-commerce/cp/settings';
+			$event->rules['mailchimp-commerce/purge'] = 'mailchimp-commerce/cp/purge';
 		}
 
 		$event->rules['mailchimp-commerce/synced/products'] = 'mailchimp-commerce/synced/products';